# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hark HTTP API",
    "version": "1",
    "summary": "Send iOS notifications, approval prompts, and Live Activities from webhooks and API clients.",
    "description": "OpenAPI definition for Hark. Each deployment has one admin who provisions regular user accounts. Sessions provide account access, scoped API tokens provide software access, and webhook tokens identify services. Account provisioning requires an admin session. See /docs.md for detailed behavior and examples.",
    "license": { "name": "MIT", "identifier": "MIT" }
  },
...
  "servers": [{ "url": "/", "description": "This Hark deployment" }],
...
  "paths": {
...
    "/auth/password": {
      "post": {
        "tags": ["Authentication"], "summary": "Change the owner password", "operationId": "changePassword",
        "security": [{ "sessionBearer": [] }, { "sessionCookie": [] }], "x-hark-auth": "session",
        "requestBody": { "$ref": "#/components/requestBodies/ChangePassword" },
        "responses": { "204": { "description": "Password changed; other sessions retired." }, "default": { "$ref": "#/components/responses/Error" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "A harksess_ session or a scoped hark_ API token, as allowed by the operation." },
      "sessionBearer": { "type": "http", "scheme": "bearer", "description": "The account owner's harksess_ session token." },
      "apiToken": { "type": "http", "scheme": "bearer", "description": "A scoped hark_ API token. Required scopes are listed in x-hark-scopes." },
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "hark_session", "description": "The session cookie. HTTPS deployments use the name __Host-hark_session." }
    },
...
    "requestBodies": {
...
      "ChangePassword": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" } } } },
...
    },
    "responses": {
      "Error": { "description": "Error response. Branch on error.code; message is not stable.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
    },
    "schemas": {
...
      "FieldError": {
        "type": "object", "additionalProperties": false, "required": ["field", "message"],
        "properties": { "field": { "type": "string" }, "message": { "type": "string" } }
      },
      "Error": {
        "type": "object", "additionalProperties": false, "required": ["code", "message"],
        "properties": {
          "code": { "type": "string", "description": "Stable machine-readable code." },
          "message": { "type": "string", "description": "Human-readable, unstable English message." },
          "fields": { "type": "array", "items": { "$ref": "#/components/schemas/FieldError" } }
        }
      },
      "ErrorEnvelope": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "$ref": "#/components/schemas/Error" } }
      },
...
      "ChangePasswordRequest": {
        "type": "object", "additionalProperties": false, "required": ["current_password", "new_password"],
        "properties": {
          "current_password": { "type": "string", "writeOnly": true },
          "new_password": { "type": "string", "minLength": 12, "maxLength": 256, "writeOnly": true }
        }
      },
...
    }
  }
}